ISSUE-146: Extract MRT recovery candidates helper
